Refactor get_packages into separate helper module
